+++ torch/quantization/scaled_mm.py
@@ -32,10 +32,37 @@ def scaled_mm(
     swizzle_b: SwizzleType | List[SwizzleType] = None,
     bias: Tensor = None,
     output_dtype: torch.dtype = torch.bfloat16,
-    scale_output: Tensor | List[Tensor] = None,
-    scale_recipe_output: ScalingType | List[ScalingType] = None,
     contraction_dim: List[int] = (),
-    use_fast_accum: bool = False) -> Tensor:
+    use_fast_accum: bool = False,
+    **kwargs) -> Tensor:
+
+    use_deprecated_api = kwargs.pop('use_deprecated_scaled_mm', False)
+    if len(kwargs) > 0:
+        raise RuntimeError(
+                "kwargs contains unexpected entries, ", kwargs.keys())
+
+    if use_deprecated_api:
+        def check_valid_scale_passed(scale):
+            if isinstance(scale, (list, tuple)):
+                if len(scale) > 1:
+                    raise RuntimeError(
+                            "deprecated api only accepts single scales, got", len(scale))
+                return scale[0]
+            else:
+                return scale
+
+        scale_a = check_valid_scale_passed(scale_a)
+        scale_b = check_valid_scale_passed(scale_b)
+
+        return torch._scaled_mm(
+                mat_a,
+                mat_b,
+                scale_a,
+                scale_b,
+                bias=bias,
+                scale_result=None,
+                out_dtype=output_dtype,
+                use_fast_accum=use_fast_accum)
 
     def expand_single_value(v):
         if not isinstance(v, (list, tuple)) and v is not None:
@@ -72,8 +99,6 @@ def scaled_mm(
             enum_list_as_int_list(list_or_empty(swizzle_b)),
             bias,
             output_dtype,
-            list_or_empty(scale_output),
-            enum_list_as_int_list(list_or_empty(scale_recipe_output)),
             contraction_dim,
             use_fast_accum)
